backend/requirements.txt
Django==5.2.11
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
pillow==12.1.1
PyJWT==2.11.0
pyodbc==5.3.0
sqlparse==0.5.5
mssql-django==1.6
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
statsmodels>=0.14.0
gunicorn==23.0.0
whitenoise==6.8.2
